// contracts/generated/typescript/models/WikiComponentVersion.ts
export interface WikiComponentVersion {
    /**
     *
     * @type {string}
     * @memberof WikiComponentVersion
     */
    componentId: string;
    /**
     *
     * @type {number}
     * @memberof WikiComponentVersion
     */
    version: number;
    /**
     *
     * @type {{ [key: string]: any; }}
     * @memberof WikiComponentVersion
     */
    propsSchema: { [key: string]: any; };
    /**
     *
     * @type {WikiComponentVersionRendererRefEnum}
     * @memberof WikiComponentVersion
     */
    rendererRef: WikiComponentVersionRendererRefEnum;
    /**
     *
     * @type {WikiComponentVersionStatusEnum}
     * @memberof WikiComponentVersion
     */
    status: WikiComponentVersionStatusEnum;
    /**
     *
     * @type {string}
     * @memberof WikiComponentVersion
     */
    createdBy: string;
    /**
     *
     * @type {Date}
     * @memberof WikiComponentVersion
     */
    createdAt: Date;
    /**
     *
     * @type {Date}
     * @memberof WikiComponentVersion
     */
    publishedAt?: Date;
}


/**
 * @export
 */
export const WikiComponentVersionRendererRefEnum = {
    BuiltinKeyValue: 'builtin.key_value',
    BuiltinEntityClaimInfobox: 'builtin.entity_claim_infobox'
} as const;
export type WikiComponentVersionRendererRefEnum = typeof WikiComponentVersionRendererRefEnum[keyof typeof WikiComponentVersionRendererRefEnum];

/**
 * @export
 */
export const WikiComponentVersionStatusEnum = {
    Draft: 'draft',
    Published: 'published',
    Deprecated: 'deprecated'
} as const;
export type WikiComponentVersionStatusEnum = typeof WikiComponentVersionStatusEnum[keyof typeof WikiComponentVersionStatusEnum];