// contracts/generated/typescript/models/WriteWikiComponentVersionRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Anby Wiki API
 * Anby Wiki HTTP API 权威契约（OpenAPI 3.1）。 本文件是所有接口行为的唯一权威定义：错误模型、分页、幂等头与请求 ID 在此统一定义。 前端客户端由 OpenAPI Generator 从本文件生成，禁止手写 URL/DTO。 Go 侧 handler 手写实现，由 Contract Test 保证与本文件一致。
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface WriteWikiComponentVersionRequest
 */
export interface WriteWikiComponentVersionRequest {
    /**
     * 根 type 必须为 object 的 JSON Schema。
     * @type {{ [key: string]: any; }}
     * @memberof WriteWikiComponentVersionRequest
     */
    propsSchema: { [key: string]: any; };
    /**
     *
     * @type {WriteWikiComponentVersionRequestRendererRefEnum}
     * @memberof WriteWikiComponentVersionRequest
     */
    rendererRef: WriteWikiComponentVersionRequestRendererRefEnum;
}


/**
 * @export
 */
export const WriteWikiComponentVersionRequestRendererRefEnum = {
    BuiltinKeyValue: 'builtin.key_value',
    BuiltinEntityClaimInfobox: 'builtin.entity_claim_infobox'
} as const;
export type WriteWikiComponentVersionRequestRendererRefEnum = typeof WriteWikiComponentVersionRequestRendererRefEnum[keyof typeof WriteWikiComponentVersionRequestRendererRefEnum];


/**
 * Check if a given object implements the WriteWikiComponentVersionRequest interface.
 */
export function instanceOfWriteWikiComponentVersionRequest(value: object): value is WriteWikiComponentVersionRequest {
    if ((!('propsSchema' in (value as Record<string, any>)) && !('props_schema' in (value as Record<string, any>))) || ((value as Record<string, any>)['propsSchema'] === undefined && (value as Record<string, any>)['props_schema'] === undefined)) return false;
    if ((!('rendererRef' in (value as Record<string, any>)) && !('renderer_ref' in (value as Record<string, any>))) || ((value as Record<string, any>)['rendererRef'] === undefined && (value as Record<string, any>)['renderer_ref'] === undefined)) return false;
    return true;
}

export function WriteWikiComponentVersionRequestFromJSON(json: any): WriteWikiComponentVersionRequest {
    return WriteWikiComponentVersionRequestFromJSONTyped(json, false);
}

export function WriteWikiComponentVersionRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): WriteWikiComponentVersionRequest {
    if (json == null) {
        return json;
    }
    return {

        'propsSchema': json['props_schema'],
        'rendererRef': json['renderer_ref'],
    };
}

export function WriteWikiComponentVersionRequestToJSON(json: any): WriteWikiComponentVersionRequest {
    return WriteWikiComponentVersionRequestToJSONTyped(json, false);
}

export function WriteWikiComponentVersionRequestToJSONTyped(value?: WriteWikiComponentVersionRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'props_schema': value['propsSchema'],
        'renderer_ref': value['rendererRef'],
    };
}